Make temporal extension work as EXTENSION or not.
This is necessary both for supporting versions of PostgreSQL that
don't have EXTENSION support, and for the test infrastructure to work.

The .sql.in file is used both to create the current extension version
.sql file and to generate a .sql file. However, the .sql.in file and
the generated .sql file need to live in the top level directory, or
the MODULE_PATHNAME is not replaced properly.